requirements.txt
Flask
requests
Werkzeug
flask_wtf
wtforms
google-cloud-secret-manager
google-cloud-logging

